requirements.txt
openai>=1.3.0
python-dotenv>=1.0.1
requests>=2.31.0
msal>=1.26.0
tenacity>=8.2.3
python-docx>=1.1.0
flask>=2.3.0
flask-cors>=4.0.0